src/requirements.txt
Flask==2.1.2
google-cloud-bigtable==2.10.0
google-cloud-core==2.3.0